fix only not deallocated in current time range
not_deallocated_ fields designate the segment of the time line. We are
interested in the question whether our deallocation point falls outside
of this segment.

This means we *have* to check both the segments to filter the point out,
so the current code which sequentially checks both ends can't work.

Let's write a match to more intuitively explain what we want here.

Basically, we want to forbid cases where `l, d, r` are in that order
(that is, `d` between `l` and `r`), so just code that up.

This is a follow up to https://github.com/koute/bytehound/commit/e981ea352c3a0dce25bb733ddd48e00549755d0d